backend/lockers/api/urls.py
from django.urls import path
from rest_framework.response import Response
from rest_framework.views import APIView

class HealthView(APIView):
    def get(self, request):
        return Response({"status": "ok"})

urlpatterns = [
    path("health", HealthView.as_view(), name="health"),
]