backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
drf-spectacular>=0.27
django-cors-headers>=4.3
psycopg[binary]>=3.1
httpx>=0.27
python-dotenv>=1.0
pytest>=8.0
pytest-django>=4.7
ruff>=0.4